ad73985e75684cb4ac4dadb9d3d86ad0d66612a0

+++ modules/audio_processing/aec3/filter_analyzer.h
@@ -12,6 +12,7 @@
 #define MODULES_AUDIO_PROCESSING_AEC3_FILTER_ANALYZER_H_
 
 #include <stddef.h>
+
 #include <array>
 #include <memory>
 #include <vector>
@@ -50,7 +51,7 @@ class FilterAnalyzer {
   float Gain() const { return gain_; }
 
   // Returns the number of blocks for the current used filter.
-  float FilterLengthBlocks() const { return filter_length_blocks_; }
+  int FilterLengthBlocks() const { return filter_length_blocks_; }
 
   // Returns the preprocessed filter.
   rtc::ArrayView<const float> GetAdjustedFilter() const { return h_highpass_; }
